requirements.txt
# Programming
tqdm
matplotlib

# Preprocess data
pillow
pandas

# Deep learning
tensorflow==2.4.0